Global Light Changes + Options Menu + Firewood Fix (#72)
* Part I - Functional Base

* Part I - Switching Position Holders

* Part I - End Of Version 1

Path vertices behave as intended, but freeform light does not. Part II will rework light propagation for more elegant behavior.

* Part II - Basic Version Set-Up

Includes Freeform Lights for each room + The basic (T-T) version of the room light-up mechanic.

* Part 3 - Scene Patcher & Receding Light Implementation

If my Skryim Mod Development shenanigans have taught me anything, is that people really like compatibility patches *-*

* Quick Fix to Old Script

I overwrote the Firewood Script again at some point (can't remember why) so I am changing another script that seemed to conflict with it now ._.

* Patcher Update to support the 109th commit

Prefabs... Need... Serialization Support...
2022.3... Do I miss you? T-T

* Firewood Fix (Plate Included)

Fixed some erratic behavior introduced by the old script in A2 - 3. Since the operations are controlled dynamically, the patch should be compatible with the new 3x2 puzzle (as long as the firewood header names stay the same).

* New Options Menu

* Pause Menu Is Alive! - Part I

A relatively stable and functional version of the pause menu. I want to push this and keep it somewhere safe; I'm about to move the floor a little and  need a version to roll back to lel;

* Pause Menu Is Alive! - Intermezzo

Throwing in a testing scene + making a minimal change to the UI prefab;

* Pause Menu Is Alive! - Part II

OK, the code surgery is over and the Pause Menu still works as intended. Playtested for a bit, fixed a number of bugs, and added the final core animations in. Almost done now;

* Pause Menu Is Alive - Part III

The last mechanical part of the update. After adding SFXs it'll be good to go.

* Trying to Fix Merge Conflict - First Try

* Revert "Trying to Fix Merge Conflict - First Try"

This reverts commit 3eddc8cc26fedd92bd2c0097b487c41dbb91e8e4.

* Trying to Fix Merge Conflict(s) - Second Try

* Trying to Fix Merge Conflict(s) - Third Try

Almost there!

* Trying to Fix Merge Conflict(s) - Reapplying Changes

* Revert "Trying to Fix Merge Conflict(s) - Reapplying Changes"

This reverts commit 0bbcda91633ebab6e2c5fa714f49537477d76a4e.